refactor: AddRestaurantModal 파일 위치 이동
- UI/modal/ 에서 UI/pages/restaurant/components 로 이동

- 리뷰어 리뷰 반영 (https://github.com/woowacourse/javascript-lunch/pull/209#issuecomment-2715135434)
