+++ src/main.js
@@ -2,7 +2,7 @@ import { DOM } from './dom.js';
 import { getRestaurantList } from './Domain/services/RestaurantService.js';
 import Header from './UI/components/header/Header.js';
 import RestaurantItem from './UI/components/restaurant/RestaurantItem.js';
-import AddRestaurantModal from './UI/modal/AddRestaurantModal.js';
+import AddRestaurantModal from './UI/pages/restaurant/components/AddRestaurantModal.js';
 
 const addRestaurantModal = new AddRestaurantModal();
 new Header(() => addRestaurantModal.handleToggleModal());
